services/backend/requirements.txt
Flask>=1.1.2
flask-bcrypt>=0.7.1
flask-cors>=3.0.9
flask-restx>=0.2.0
Flask-SQLAlchemy>=2.4.4
gunicorn>=20.0.4
psycopg2-binary>=2.8.6
pyjwt>=1.7.1
python-dateutil>=2.8.1
flask-praetorian>=1.1.0
